Create the tables and associations for challenge.
This was harder than I expected. The multiple levels of joins are easier
to do in SQL than it was in the ActiveRecord syntax (and I don't
consider myself particularly good in SQL).

Although the documentation is pretty extensive, I found some annoying
holes that were filled in by suggestions in the error messages.